[FIX] Handled bad read and Segfaults on vector accesses

+++ lib/utils/entry.hh
@@ -16,6 +16,6 @@ public:
 
     json serialize_message() const;
 
-    size_t term_;
+    int term_;
     std::string command_;
 };
