+++ src/rpc/append-entries.hh
@@ -12,10 +12,10 @@ class AppendEntries : public RPC
 public:
     AppendEntries(int term,
                   size_t leader_id,
-                  size_t prev_log_index,
+                  int prev_log_index,
                   int prev_log_term,
                   std::vector<Entry> entries,
-                  size_t leader_commit);
+                  int leader_commit);
     AppendEntries(int term, const json& serialized_json);
     AppendEntries(int term, const std::string& serialized);
 
@@ -24,9 +24,9 @@ public:
     // so follower can redirect clients
     const size_t leader_id_;
     // index of log entry immediately preceding new ones
-    const size_t prev_log_index_;
+    const int prev_log_index_;
     // term of prev_log_index entry
-    const size_t prev_log_term_;
+    const int prev_log_term_;
     // log entries to store (empty for heartbeat may send more than one for efficiency)
     std::vector<Entry> entries_;
     // leader's commit_index
